Add pangram to body; commit to src from parent
The body of the text was updated to include the pangram output from autotext.tex. Autotext.tex was modified to put 25 instances per paragraph. The change in the file is being committed from the parent directory. Will it be pushed to the dependent repsoitory? or will it be kept in the parent repsoitory?